+++ CHANGELOG.md
@@ -139,6 +139,7 @@ Up requirements to Django 3.2+, Twisted 21+
 - Add support for `$dbref()` and `$search` when assigning an Attribute value
   with the `set` command. This allows assigning real objects from in-game.
 - Add ability to examine `/script` and `/channel` entities  with `examine` command.
+- Homogenize manager search methods to return querysets and not lists.
 
 
 ### Evennia 0.9.5 (2019-2020)